+++ ui/packages/shared/profile/src/ProfileSelector/index.tsx
@@ -119,7 +119,7 @@ const ProfileSelector = ({
   onSearchHook,
 }: ProfileSelectorProps): JSX.Element => {
   const {heightStyle} = useMetricsGraphDimensions(comparing, false);
-  const {viewComponent} = useParcaContext();
+  const {viewComponent, additionalMetricsGraph} = useParcaContext();
   const [queryBrowserMode, setQueryBrowserMode] = useURLState('query_browser_mode');
   const batchUpdates = useURLStateBatch();
 
@@ -290,7 +290,15 @@ const ProfileSelector = ({
 
   return (
     <>
-      <div className="mb-2 flex">
+      <div className="mb-2 flex flex-col">
+        {additionalMetricsGraph?.({
+          querySelection,
+          queryClient,
+          suffix,
+          timeRange: timeRangeSelection,
+          onTimeRangeChange: handleTimeRangeChange,
+          commitTimeRange: () => setQueryExpression(true),
+        })}
         <LabelsQueryProvider
           setMatchersString={setMatchersString}
           runQuery={setQueryExpression}