+++ ui/packages/shared/profile/src/index.tsx
@@ -16,10 +16,12 @@ import * as lz4 from 'lz4js';
 
 import type {ParamPreferences} from '@parca/components';
 
+import MatchersInput from './MatchersInput';
 import MetricsGraph, {type ContextMenuItemOrSubmenu, type Series} from './MetricsGraph';
 import ProfileExplorer from './ProfileExplorer';
 import ProfileTypeSelector from './ProfileTypeSelector';
 import {SelectWithRefresh} from './SelectWithRefresh';
+import SimpleMatchers from './SimpleMatchers';
 import CustomSelect from './SimpleMatchers/Select';
 import {
   LabelsQueryProvider,
@@ -80,6 +82,8 @@ export {
   SelectWithRefresh,
   useLabelNames,
   MetricsGraph,
+  SimpleMatchers,
+  MatchersInput,
   type ContextMenuItemOrSubmenu,
   type Series,
   LabelsQueryProvider,