+++ ui/packages/shared/utilities/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.124 (2026-02-12)
+
+**Note:** Version bump only for package @parca/utilities
+
 ## 0.0.123 (2026-02-06)
 
 **Note:** Version bump only for package @parca/utilities